V3

+++ mods/pvp/dropondie/init.lua
@@ -56,5 +56,7 @@ local function drop_all(player)
 	drop_list(player, pos, inv, "craft")
 end
 
-minetest.register_on_dieplayer(drop_all)
-minetest.register_on_leaveplayer(drop_all)
+if ctf_core.settings.server_mode ~= "mapedit" then
+	minetest.register_on_dieplayer(drop_all)
+	minetest.register_on_leaveplayer(drop_all)
+end
